backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

opencv-python
pillow
python-doctr
python-doctr[torch]
easyocr
pytesseract
scipy
numpy
scikit-image
python-bidi
PyYAML
Shapely
pyclipper
torch==2.7.0+cpu
torchvision